requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
pytest>=7.3.1
coverage>=7.2.7
requests>=2.31.0
djangorestframework-simplejwt>=5.3.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-filter>=23.5
channels>=4.0.0
daphne>=4.0.0
django-debug-toolbar>=4.3.0
django-environ>=0.11.2
gunicorn>=21.2.0
django-redis>=5.4.0
pyotp>=2.9.0
qrcode>=7.4.2
django-storages>=1.14.2
django-csp>=3.7
django-ratelimit>=4.1.0
django-axes>=6.3.0
django-otp>=1.3.0
django-two-factor-auth>=1.15.5
django-allauth>=0.60.1
django-rest-auth>=0.9.5
django-rest-knox>=4.2.0
django-cors-headers>=4.3.1
django-guardian>=2.4.0
django-ratelimit-backend>=1.0.0
django-ses>=3.5.0